+++ app/src/free/java/dev/dworks/apps/anexplorer/directory/NativeAdViewHolder.java
