+++ app/src/free/res/layout/item_doc_ad_list.xml
