+++ app/src/main/java/dev/dworks/apps/anexplorer/fragment/DirectoryFragment.java
@@ -31,7 +31,6 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Parcelable;
 import android.provider.Settings;
-import android.text.TextUtils;
 import android.util.Log;
 import android.util.SparseArray;
 import android.util.SparseBooleanArray;
@@ -55,10 +54,6 @@ import androidx.loader.app.LoaderManager.LoaderCallbacks;
 import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.google.android.gms.ads.AdListener;
-import com.google.android.gms.ads.AdLoader;
-import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.formats.UnifiedNativeAd;
 import com.google.android.material.snackbar.Snackbar;
 
 import java.util.ArrayList;
@@ -70,9 +65,7 @@ import dev.dworks.apps.anexplorer.DocumentsActivity;
 import dev.dworks.apps.anexplorer.DocumentsApplication;
 import dev.dworks.apps.anexplorer.R;
 import dev.dworks.apps.anexplorer.common.DialogBuilder;
-import dev.dworks.apps.anexplorer.common.RecyclerFragment;
-import dev.dworks.apps.anexplorer.cursor.MatrixCursor;
-import dev.dworks.apps.anexplorer.cursor.RootCursorWrapper;
+import dev.dworks.apps.anexplorer.directory.DirectoryFragmentFlavour;
 import dev.dworks.apps.anexplorer.directory.DividerItemDecoration;
 import dev.dworks.apps.anexplorer.directory.DocumentsAdapter;
 import dev.dworks.apps.anexplorer.directory.FolderSizeAsyncTask;
@@ -118,7 +111,6 @@ import static dev.dworks.apps.anexplorer.BaseActivity.TAG;
 import static dev.dworks.apps.anexplorer.DocumentsApplication.isSpecialDevice;
 import static dev.dworks.apps.anexplorer.DocumentsApplication.isTelevision;
 import static dev.dworks.apps.anexplorer.DocumentsApplication.isWatch;
-import static dev.dworks.apps.anexplorer.directory.NativeAdViewHolder.BUNDLE_AD_KEY;
 import static dev.dworks.apps.anexplorer.misc.AnalyticsManager.FILE_COUNT;
 import static dev.dworks.apps.anexplorer.misc.AnalyticsManager.FILE_MOVE;
 import static dev.dworks.apps.anexplorer.misc.AnalyticsManager.FILE_TYPE;
@@ -131,11 +123,8 @@ import static dev.dworks.apps.anexplorer.misc.Utils.EXTRA_IGNORE_STATE;
 import static dev.dworks.apps.anexplorer.misc.Utils.EXTRA_QUERY;
 import static dev.dworks.apps.anexplorer.misc.Utils.EXTRA_ROOT;
 import static dev.dworks.apps.anexplorer.misc.Utils.EXTRA_TYPE;
-import static dev.dworks.apps.anexplorer.misc.Utils.NATIVE_APP_UNIT_ID;
-import static dev.dworks.apps.anexplorer.misc.Utils.NATIVE_BIG_APP_UNIT_ID;
 import static dev.dworks.apps.anexplorer.misc.Utils.isRooted;
 import static dev.dworks.apps.anexplorer.model.DocumentInfo.getCursorInt;
-import static dev.dworks.apps.anexplorer.model.DocumentInfo.getCursorLong;
 import static dev.dworks.apps.anexplorer.model.DocumentInfo.getCursorString;
 import static dev.dworks.apps.anexplorer.ui.RecyclerViewPlus.TYPE_GRID;
 import static dev.dworks.apps.anexplorer.ui.RecyclerViewPlus.TYPE_LIST;
@@ -143,7 +132,7 @@ import static dev.dworks.apps.anexplorer.ui.RecyclerViewPlus.TYPE_LIST;
 /**
  * Display the documents inside a single directory.
  */
-public class DirectoryFragment extends RecyclerFragment implements MenuItem.OnMenuItemClickListener {
+public class DirectoryFragment extends DirectoryFragmentFlavour implements MenuItem.OnMenuItemClickListener {
 
 	private static final String KEY_ADAPTER = "key_adapter";
 	private CompatTextView mEmptyView;
@@ -156,8 +145,6 @@ public class DirectoryFragment extends RecyclerFragment implements MenuItem.OnMe
 	public static final int ANIM_SIDE = 2;
 	public static final int ANIM_DOWN = 3;
 	public static final int ANIM_UP = 4;
-
-	public static final int AD_POSITION = 5;
 
 	private int mType = TYPE_NORMAL;
 	private String mStateKey;
@@ -190,8 +177,6 @@ public class DirectoryFragment extends RecyclerFragment implements MenuItem.OnMe
 	private IconHelper mIconHelper;
 	private MultiChoiceHelper mMultiChoiceHelper;
 	boolean selectAll = true;
-	private AdLoader adLoader;
-	private ArrayList<UnifiedNativeAd> mNativeAds = new ArrayList<>();
 
 	public static void showNormal(FragmentManager fm, RootInfo root, DocumentInfo doc, int anim) {
 		show(fm, TYPE_NORMAL, root, doc, null, anim);
@@ -279,6 +264,7 @@ public class DirectoryFragment extends RecyclerFragment implements MenuItem.OnMe
 
 		final DocumentsActivity context = (DocumentsActivity)getActivity();
 		final State state = getDisplayState(DirectoryFragment.this);
+		final boolean showNativeAds = !context.showBannerAds;
 
 		root = getArguments().getParcelable(EXTRA_ROOT);
 		doc = getArguments().getParcelable(EXTRA_DOC);
@@ -351,7 +337,7 @@ public class DirectoryFragment extends RecyclerFragment implements MenuItem.OnMe
 				}
 
 				boolean showData = mType == TYPE_RECENT_OPEN || AppPaymentFlavour.isPurchased()
-						|| isSpecialDevice();
+						|| isSpecialDevice() || !showNativeAds;
 				if (showData) {
 					showData(result);
 				} else {
@@ -375,86 +361,7 @@ public class DirectoryFragment extends RecyclerFragment implements MenuItem.OnMe
 		updateDisplayState();
 	}
 
-	private void loadNativeAds(final DirectoryResult result) {
-		int cursorCount = result.cursor != null ? result.cursor.getCount() : 0;
-		if(cursorCount <= 5){
-			showData(result);
-			return;
-		}
-		String appUnitId = result.mode == MODE_GRID ? NATIVE_BIG_APP_UNIT_ID : NATIVE_APP_UNIT_ID;
-		AdLoader.Builder builder = new AdLoader.Builder(getActivity(), appUnitId);
-		mNativeAds = new ArrayList<>();
-		adLoader = builder.forUnifiedNativeAd(new UnifiedNativeAd.OnUnifiedNativeAdLoadedListener() {
-			@Override
-			public void onUnifiedNativeAdLoaded(UnifiedNativeAd unifiedNativeAd) {
-				mNativeAds.add(unifiedNativeAd);
-				if (null != adLoader && !adLoader.isLoading()) {
-					insertNativeAds(result);
-				}
-			}
-		}).withAdListener(new AdListener() {
-			@Override
-			public void onAdFailedToLoad(int errorCode) {
-				if (null != adLoader && !adLoader.isLoading()) {
-					insertNativeAds(result);
-				}
-			}
-		}).build();
-
-		boolean dontLoadAds = null == adLoader;
-		if (dontLoadAds){
-			showData(result);
-			return;
-		}
-
-		int numberOfAds = cursorCount / AD_POSITION;
-		adLoader.loadAds(new AdRequest.Builder().build(), numberOfAds);
-	}
-
-	private void insertNativeAds(DirectoryResult result) {
-		int cursorCount = result.cursor != null ? result.cursor.getCount() : 0;
-		int adsCount = mNativeAds.size();
-		if (adsCount <= 0 || cursorCount <= AD_POSITION) {
-			return;
-		}
-		int offset = (cursorCount / mNativeAds.size()) + 1;
-		int index = 0;
-		MatrixCursor matrixCursor = new MatrixCursor(result.cursor.getColumnNames());
-		Bundle bundle = new Bundle();
-		bundle.putSerializable(BUNDLE_AD_KEY, mNativeAds);
-		matrixCursor.respond(bundle);
-		result.cursor.moveToPosition(-1);
-		for (int i = 0; i < (cursorCount + adsCount); i++) {
-			if (i % offset == 0) {
-				final MatrixCursor.RowBuilder row = matrixCursor.newRow();
-				row.add(RootCursorWrapper.COLUMN_AUTHORITY, "");
-				row.add(Document.COLUMN_DOCUMENT_ID, "");
-				row.add(Document.COLUMN_FLAGS, index);
-				index++;
-			} else {
-				result.cursor.moveToNext();
-				String authority =  getCursorString(result.cursor, RootCursorWrapper.COLUMN_AUTHORITY);
-				if(TextUtils.isEmpty(authority)){
-					continue;
-				}
-				final MatrixCursor.RowBuilder row = matrixCursor.newRow();
-				row.add(RootCursorWrapper.COLUMN_AUTHORITY, authority);
-				row.add(Document.COLUMN_DOCUMENT_ID, getCursorString(result.cursor, Document.COLUMN_DOCUMENT_ID));
-				row.add(Document.COLUMN_MIME_TYPE, getCursorString(result.cursor, Document.COLUMN_MIME_TYPE));
-				row.add(Document.COLUMN_DISPLAY_NAME, getCursorString(result.cursor, Document.COLUMN_DISPLAY_NAME));
-				row.add(Document.COLUMN_LAST_MODIFIED, getCursorLong(result.cursor, Document.COLUMN_LAST_MODIFIED));
-				row.add(Document.COLUMN_FLAGS, getCursorInt(result.cursor, Document.COLUMN_FLAGS));
-				row.add(Document.COLUMN_SUMMARY, getCursorString(result.cursor, Document.COLUMN_SUMMARY));
-				row.add(Document.COLUMN_SIZE, getCursorLong(result.cursor, Document.COLUMN_SIZE));
-				row.add(Document.COLUMN_ICON, getCursorInt(result.cursor, Document.COLUMN_ICON));
-				row.add(Document.COLUMN_PATH, getCursorString(result.cursor, Document.COLUMN_PATH));
-			}
-		}
-		adLoader = null;
-		result.cursor = matrixCursor;
-		showData(result);
-	}
-
+	@Override
 	public void showData(DirectoryResult result){
 
 		if (!isAdded())
